Refactor labeling dialog into its own component
