+++ wsi_superpixel_guided_labeling/web_client/views/vue/components/ActiveLearning/ActiveLearningKeyboardShortcuts.vue
@@ -3,7 +3,7 @@ import Vue from 'vue';
 
 import _ from 'underscore';
 
-import { comboHotkeys } from './constants';
+import { comboHotkeys } from '../constants.js';
 import { store, previousCard, nextCard, assignHotkey } from '../store.js';
 
 import MouseAndKeyboardControls from '../MouseAndKeyboardControls.vue';
