+++ wsi_superpixel_guided_labeling/web_client/views/vue/components/ActiveLearningSetup/ActiveLearningInitialLabels.vue
@@ -4,64 +4,53 @@ import Vue from 'vue';
 import _ from 'underscore';
 import { restRequest } from '@girder/core/rest';
 import { ViewerWidget } from '@girder/large_image_annotation/views';
-import ColorPickerInput from '@girder/histomicsui/vue/components/ColorPickerInput.vue';
-import { confirm } from '@girder/core/dialog';
 
-import ActiveLearningMergeConfirmation from './ActiveLearningMergeConfirmation.vue';
+import ActiveLearningLabeling from './ActiveLearningLabeling.vue';
 import AnnotationOpacityControl from '../AnnotationOpacityControl.vue';
 import MouseAndKeyboardControls from '../MouseAndKeyboardControls.vue';
-import { comboHotkeys, schemeTableau10 } from '../ActiveLearning/constants.js';
-import { store, updatePixelmapLayerStyle } from '../store.js';
+import { comboHotkeys, boundaryColor } from '../constants.js';
+import { store, updatePixelmapLayerStyle, getFillColor } from '../store.js';
 
 // Define some helpful constants for adding categories
-const boundaryColor = 'rgba(0, 0, 0, 1)';
 const defaultCategory = {
     label: 'default',
     fillColor: 'rgba(0, 0, 0, 0)',
     strokeColor: 'rgba(0, 0, 0, 1)'
 };
-const colorPattern = /^(#[0-9a-fA-F]{3,4}|#[0-9a-fA-F]{6}|#[0-9a-fA-F]{8}|rgb\(\d+,\s*\d+,\s*\d+\)|rgba\(\d+,\s*\d+,\s*\d+,\s*(\d?\.|)\d+\))$/;
 
 export default Vue.extend({
     components: {
-        ColorPickerInput,
         MouseAndKeyboardControls,
         AnnotationOpacityControl,
-        ActiveLearningMergeConfirmation
+        ActiveLearningLabeling
     },
     props: ['backboneParent', 'imageNamesById', 'annotationsByImageId', 'availableImages', 'categoryMap'],
     data() {
         return {
             hasLoaded: false,
-            showLabelingContainer: true,
             showInfoContainer: true,
-            editing: -1,
 
             // data tracking current categories/currently active category
-            categories: [],
             categoryIndex: 0,
-            currentCategoryLabel: 'New Category',
-            currentCategoryFillColor: this.getFillColor(0),
-            checkedCategories: [],
 
             // keep track of the current image and annotation to edit
-            currentImageId: '',
             superpixelAnnotation: null,
             superpixelElement: null,
-            newImagesAvailable: false,
 
             // data to track the viewer widget/map/layers if needed
             viewerWidget: null,
             overlayLayer: null,
             pixelmapRendered: false,
-            pixelmapPaintValue: null,
-            pixelmapPaintBrush: false
+            pixelmapPaintValue: null
         };
     },
     computed: {
+        categoriesAndIndices() {
+            return store.categoriesAndIndices;
+        },
         currentFormErrors() {
             const errors = [];
-            const categoryNames = _.map(this.categories, (category) => category.category.label);
+            const categoryNames = _.map(this.categoriesAndIndices, (category) => category.category.label);
             const differentCategoryNames = new Set(categoryNames);
             if (categoryNames.length !== differentCategoryNames.size) {
                 errors.push('All categories must have unique names.');
@@ -71,27 +60,16 @@ export default Vue.extend({
         currentCategoryFormValid() {
             return this.currentFormErrors.length === 0;
         },
-        currentLabelingErrors() {
-            const errors = [];
-            const counts = _.map(Object.keys(this.labeledSuperpixelCounts), (entry) => this.labeledSuperpixelCounts[entry].count);
-            if (_.filter(counts, (count) => count > 0).length < 2) {
-                errors.push('You must label superpixels for at least two different categories.');
-            }
-            return errors;
-        },
-        currentLabelsValid() {
-            return this.currentLabelingErrors.length === 0;
-        },
         usingBoundaries() {
             return this.superpixelElement.boundaries;
         },
         allNewCategories() {
-            const activeLearningCategories = _.map(this.categories, (category) => category.category);
+            const activeLearningCategories = _.map(this.categoriesAndIndices, (category) => category.category);
             return [defaultCategory, ...activeLearningCategories];
         },
         labeledSuperpixelCounts() {
             const counts = {};
-            _.forEach(this.categories, (categoryAndIndices, index) => {
+            _.forEach(this.categoriesAndIndices, (categoryAndIndices, index) => {
                 const label = categoryAndIndices.category.label;
                 const fillColor = categoryAndIndices.category.fillColor;
                 const key = `${index}_${label}`;
@@ -112,35 +90,18 @@ export default Vue.extend({
         hotkeys() {
             return store.hotkeys;
         },
-        selectedLabels() {
-            return new Map(_.map(this.checkedCategories, (i) => {
-                const label = this.categories[i].category.label;
-                return [i, this.labeledSuperpixelCounts[`${i}_${label}`]];
-            }));
+        pixelmapPaintBrush() {
+            return store.pixelmapPaintBrush;
+        },
+        currentImageId() {
+            return store.currentImageId;
         }
     },
     watch: {
-        currentCategoryLabel(newLabel, oldLabel) {
-            if (newLabel === oldLabel) {
-                return;
-            }
-            this.categories[this.categoryIndex].category.label = this.currentCategoryLabel;
-            this.synchronizeCategories();
-        },
-        currentCategoryFillColor(newColor, oldColor) {
-            if (newColor === oldColor || !colorPattern.test(newColor)) {
-                return;
-            }
-            this.categories[this.categoryIndex].category.fillColor = this.currentCategoryFillColor;
-            this.synchronizeCategories();
-        },
         categoryIndex(index) {
-            if (index < 0 || index >= this.categories.length) {
+            if (index < 0 || index >= this.categoriesAndIndices.length) {
                 this.categoryIndex = 0;
-                return;
             }
-            this.currentCategoryLabel = this.categories[this.categoryIndex].category.label;
-            this.currentCategoryFillColor = this.categories[this.categoryIndex].category.fillColor;
         },
         currentImageId() {
             this.superpixelAnnotation = this.annotationsByImageId[this.currentImageId].labels;
@@ -148,28 +109,12 @@ export default Vue.extend({
         },
         pixelmapPaintBrush() {
             this.updateActionModifiers();
-        },
-        editing() {
-            if (this.editing === -1) {
-                return;
-            }
-            const key = `label_${this.editing}`;
-            this.$nextTick(() => this.$refs[key][0].focus());
-        },
-        availableImages(newAvail, oldAvail) {
-            const newImage = _.difference(newAvail, oldAvail)[0];
-            if (newImage === this.currentImageId) {
-                // Update image with annotations
-                this.superpixelAnnotation = this.annotationsByImageId[this.currentImageId].labels;
-                this.setupViewer();
-            }
-            this.newImagesAvailable = true;
         }
     },
     mounted() {
         window.addEventListener('keydown', this.keydownListener);
-        this.currentImageId = Object.keys(this.imageNamesById)[0];
-        this.superpixelAnnotation = this.annotationsByImageId[this.currentImageId].labels;
+        store.currentImageId = Object.keys(this.imageNamesById)[0];
+        this.superpixelAnnotation = this.annotationsByImageId[store.currentImageId].labels;
         store.annotationsByImageId = this.annotationsByImageId;
         store.backboneParent = this.backboneParent;
         this.setupViewer();
@@ -264,13 +209,13 @@ export default Vue.extend({
             _.forEach(Object.entries(this.annotationsByImageId), ([imageId, annotations]) => {
                 if (_.has(annotations, 'labels')) {
                     const pixelmapElement = annotations.labels.get('annotation').elements[0];
-                    const existingCategories = _.map(this.categories, (category) => category.category.label);
+                    const existingCategories = _.map(this.categoriesAndIndices, (category) => category.category.label);
                     this.createCategoriesFromPixelmapElement(pixelmapElement, imageId, existingCategories);
                 }
             });
-            if (this.categories.length === 0) {
-                const fillColor = this.getFillColor(this.categories.length);
-                this.categories.push({
+            if (this.categoriesAndIndices.length === 0) {
+                const fillColor = getFillColor(this.categoriesAndIndices.length);
+                store.categoriesAndIndices.push({
                     category: {
                         label: 'New Category',
                         fillColor,
@@ -280,8 +225,6 @@ export default Vue.extend({
                 });
             }
             this.categoryIndex = 0;
-            this.currentCategoryLabel = this.categories[0].category.label;
-            this.currentCategoryFillColor = this.categories[0].category.fillColor;
         },
         /**
          * Given a pixelmap annotation element and a list of categories, work to populate the component's
@@ -308,13 +251,13 @@ export default Vue.extend({
                     if (!existingCategories.includes(category.label)) {
                         const indices = {};
                         indices[this.currentImageId] = labeledSuperpixels;
-                        this.categories.push({
+                        store.categoriesAndIndices.push({
                             category,
                             indices
                         });
                     } else if (labeledSuperpixels.size) {
-                        const categoryToUpdateIndex = _.findIndex(this.categories, (categoryAndIndices) => categoryAndIndices.category.label === category.label);
-                        this.categories[categoryToUpdateIndex].indices[imageId] = labeledSuperpixels;
+                        const categoryToUpdateIndex = _.findIndex(this.categoriesAndIndices, (categoryAndIndices) => categoryAndIndices.category.label === category.label);
+                        store.categoriesAndIndices[categoryToUpdateIndex].indices[imageId] = labeledSuperpixels;
                     }
                 }
             });
@@ -400,115 +343,21 @@ export default Vue.extend({
         },
         updateRunningLabelCounts(boundaries, index, newLabel, oldLabel) {
             const elementValueIndex = boundaries ? index / 2 : index;
-            const currentCategoryIndices = this.categories[this.categoryIndex].indices[this.currentImageId] || new Set();
+            const currentCategoryIndices = this.categoriesAndIndices[this.categoryIndex].indices[this.currentImageId] || new Set();
             if (!currentCategoryIndices.size) {
-                this.categories[this.categoryIndex].indices[this.currentImageId] = currentCategoryIndices;
+                store.categoriesAndIndices[this.categoryIndex].indices[this.currentImageId] = currentCategoryIndices;
             }
             if (newLabel === 0) {
-                this.categories[oldLabel - 1].indices[this.currentImageId].delete(elementValueIndex);
+                store.categoriesAndIndices[oldLabel - 1].indices[this.currentImageId].delete(elementValueIndex);
                 currentCategoryIndices.delete(elementValueIndex);
             } else if (oldLabel === 0) {
                 currentCategoryIndices.add(elementValueIndex);
             } else {
-                this.categories[oldLabel - 1].indices[this.currentImageId].delete(elementValueIndex);
+                store.categoriesAndIndices[oldLabel - 1].indices[this.currentImageId].delete(elementValueIndex);
                 currentCategoryIndices.add(elementValueIndex);
             }
             // Force computed properties to update
-            this.categories = [...this.categories];
-        },
-        /*************************
-         * RESPOND TO USER INPUT *
-         *************************/
-        addCategory(newName, newFillColor) {
-            if (_.isUndefined(newName)) {
-                newName = 'New Category';
-            }
-            if (_.isUndefined(newFillColor)) {
-                newFillColor = this.getFillColor(this.categories.length);
-            }
-
-            this.categories.push({
-                category: {
-                    label: this.enforceUniqueName(newName),
-                    fillColor: newFillColor,
-                    strokeColor: boundaryColor
-                },
-                indices: {}
-            });
-            this.categoryIndex = this.categories.length - 1;
-        },
-        combineCategories(indices, isMerge) {
-            // Remove the selected categories
-            indices = _.sortBy(indices, (i) => i).reverse();
-            const oldCategories = _.map(indices, (index) => {
-                return this.categories.splice(index, 1)[0];
-            });
-
-            _.forEach(Object.keys(this.annotationsByImageId), (imageId) => {
-                if (!_.has(this.annotationsByImageId[imageId], 'labels')) {
-                    return;
-                }
-                const labels = this.annotationsByImageId[imageId].labels;
-                const pixelmapElement = labels.get('annotation').elements[0];
-                pixelmapElement.categories = [...this.allNewCategories];
-
-                // Reset removed category labels to the default category
-                _.forEach(oldCategories, (category, val) => {
-                    const indices = category.indices[imageId] || new Set();
-                    _.forEach([...indices], (index) => {
-                        pixelmapElement.values[index] = 0;
-                    });
-                    if (isMerge) {
-                        // All merged indices should be assigned to the new combined category
-                        const newIndices = _.last(this.categories).indices[imageId] || new Set();
-                        _.last(this.categories).indices[imageId] = new Set([...newIndices, ...indices]);
-                    }
-                });
-
-                // Indices have shifted after removing the selected categories
-                _.forEach(this.categories, (category, val) => {
-                    const indices = category.indices[imageId] || new Set();
-                    _.forEach([...indices], (index) => {
-                        pixelmapElement.values[index] = val + 1;
-                    });
-                });
-            });
-            this.checkedCategories = [];
-            this.drawPixelmapAnnotation();
-            this.saveAnnotations(true);
-            this.updateConfig();
-        },
-        mergeCategory(newLabel, newFillColor) {
-            this.addCategory('Merged Categories', newFillColor);
-            this.combineCategories(this.checkedCategories, true);
-            _.last(this.categories).category.label = this.enforceUniqueName(newLabel);
-        },
-        deleteCategory(indices) {
-            const labelCounts = _.reduce([...this.selectedLabels.values()], (acc, selected) => {
-                return acc + selected.count;
-            }, 0);
-            if (labelCounts === 0) {
-                // If nothing was labeled we don't need a warning dialog
-                this.combineCategories(indices, false);
-                return;
-            }
-            const message = `Deleting categories cannot be undone. Are you
-                            sure you want to delete all ${labelCounts} labeled
-                            superpixels?`;
-            confirm({
-                title: 'Warning',
-                text: message,
-                yesText: 'Delete Selected',
-                confirmCallback: () => this.combineCategories(indices, false)
-            });
-        },
-        togglePicker(event, index) {
-            const picker = this.$refs.colorpicker[index];
-            const colorPicker = picker.$refs.colorPicker;
-            if (event.target.className === 'current-color' && colorPicker) {
-                // Default to th RGBA input
-                colorPicker.fieldsIndex++;
-            }
+            store.categoriesAndIndices = [...this.categoriesAndIndices];
         },
         /***********
          * UTILITY *
@@ -531,13 +380,8 @@ export default Vue.extend({
                 this.updateConfig();
             }
         },
-        getFillColor(index) {
-            const hexColor = schemeTableau10[index % 10];
-            const [r, g, b] = hexColor.slice(1).match(/.{1,2}/g);
-            return `rgba(${parseInt(r, 16)}, ${parseInt(g, 16)}, ${parseInt(b, 16)}, 0.5)`;
-        },
         enforceUniqueName(name) {
-            const existingNames = _.map(this.categories, (category) => category.category.label);
+            const existingNames = _.map(this.categoriesAndIndices, (category) => category.category.label);
             let count = 1;
             let uniqueName = name;
             while (_.some(existingNames, (en) => en.includes(uniqueName)) && count < 50) {
@@ -557,6 +401,11 @@ export default Vue.extend({
                 }
             });
         },
+        combineCategoriesHandler() {
+            this.drawPixelmapAnnotation();
+            this.saveAnnotations(true);
+            this.updateConfig();
+        },
         /**********************************
          * USE BACKBONE CONTAINER METHODS *
          **********************************/
@@ -569,255 +418,37 @@ export default Vue.extend({
         },
         updateConfig: _.debounce(function () {
             this.backboneParent.updateHistomicsYamlConfig(store.categories);
-        }, 500),
-        hotkeyFromIndex(index) {
-            return _.find([...this.hotkeys], ([, v]) => v === index)[0];
-        }
+        }, 500)
     }
 });
 </script>
 
 <template>
   <div>
-    <div :class="{'h-labeling-container': true, 'h-collapsed': !showLabelingContainer}">
-      <div class="h-container-title">
-        <button
-          class="h-collapse-button"
-          @click="showLabelingContainer = !showLabelingContainer"
-        >
-          <i
-            v-if="showLabelingContainer"
-            class="icon-angle-double-left"
-            data-toggle="tooltip"
-            title="Hide labeling editor"
-          />
-          <i
-            v-else
-            class="icon-angle-double-right"
-            data-toggle="tooltip"
-            title="Show labeling editor"
-          />
-        </button>
-        <h4 v-if="showLabelingContainer">
-          <i class="icon-tags" />
-          Labeling
-        </h4>
-        <i
-          v-else
-          class="icon-tags"
-        />
-        <div
-          v-if="pixelmapRendered"
-          class="btn-group"
-        >
-          <button
-            v-if="showLabelingContainer"
-            :class="['btn btn-default', !pixelmapPaintBrush && 'active btn-primary']"
-            data-toggle="tooltip"
-            title="Left-click + drag to pan"
-            @click="pixelmapPaintBrush = !pixelmapPaintBrush"
-          >
-            <i class="icon-move" />
-          </button>
-          <button
-            v-if="showLabelingContainer"
-            :class="['btn btn-default', pixelmapPaintBrush && 'active btn-primary']"
-            data-toggle="tooltip"
-            title="Left-click + drag to paint"
-            @click="pixelmapPaintBrush = !pixelmapPaintBrush"
-          >
-            <i class="icon-paint-roller" />
-          </button>
-        </div>
-      </div>
-      <div
-        v-if="showLabelingContainer"
-        class="h-al-setup-categories"
-      >
-        <div>
-          <div class="h-category-form">
-            <div class="h-form-controls">
-              <div>
-                <label for="currentImage">Image</label>
-                <select
-                  id="currentImage"
-                  v-model="currentImageId"
-                  :class="['h-al-image-select', newImagesAvailable && 'h-al-image-select-new']"
-                  :style="[!availableImages.includes(currentImageId) && {'font-style': 'italic'}]"
-                  @click="newImagesAvailable = false"
-                >
-                  <option
-                    v-for="imageId in Object.keys(imageNamesById)"
-                    :key="imageId"
-                    :value="imageId"
-                    :style="[!availableImages.includes(imageId) ? {'font-style': 'italic'} : {'font-style': 'normal'}]"
-                  >
-                    {{ imageNamesById[imageId] }}
-                  </option>
-                </select>
-              </div>
-            </div>
-          </div>
-          <div>
-            <table class="table">
-              <tr>
-                <th><i class="icon-keyboard" /></th>
-                <th>Label</th>
-                <th>Superpixels</th>
-              </tr>
-              <tr
-                v-for="(key, index) in Object.keys(labeledSuperpixelCounts)"
-                :key="index"
-                :class="{'h-selected-row': categoryIndex === index}"
-                @click="categoryIndex = index"
-              >
-                <td>{{ hotkeyFromIndex(index + 1) }}</td>
-                <td
-                  v-if="editing === index"
-                  class="table-labels"
-                >
-                  <input
-                    id="category-label"
-                    :ref="`label_${index}`"
-                    v-model="currentCategoryLabel"
-                    class="form-control input-sm category-input"
-                    @keyup.enter="editing = -1"
-                    @blur="editing = -1"
-                    @focus="$event.target.select()"
-                  >
-                  <button
-                    class="btn h-table-button"
-                    data-toggle="tooltip"
-                    title="Accept changes"
-                    @click="editing = -1"
-                  >
-                    <i class="icon-ok" />
-                  </button>
-                </td>
-                <td
-                  v-else
-                  class="table-labels"
-                >
-                  {{ labeledSuperpixelCounts[key].label }}
-                  <div class="editing-icons">
-                    <button
-                      class="btn h-table-button"
-                      data-toggle="tooltip"
-                      title="Edit label name"
-                      @click="editing = index"
-                    >
-                      <i class="icon-pencil" />
-                    </button>
-                  </div>
-                </td>
-                <td>{{ labeledSuperpixelCounts[key].count }}</td>
-                <td
-                  id="colorPickerInput"
-                  @click="(e) => togglePicker(e, index)"
-                >
-                  <color-picker-input
-                    :key="key"
-                    ref="colorpicker"
-                    v-model="currentCategoryFillColor"
-                    class="condensed-color-picker"
-                    :color="labeledSuperpixelCounts[key].fillColor"
-                    data-toggle="tooltip"
-                    title="Change label color"
-                  />
-                </td>
-                <td>
-                  <input
-                    v-model="checkedCategories"
-                    type="checkbox"
-                    :value="index"
-                  >
-                </td>
-              </tr>
-            </table>
-            <div class="h-remove-categories">
-              <button
-                class="btn btn-danger btn-xs"
-                :disabled="checkedCategories.length < 1"
-                data-toggle="tooltip"
-                title="Delete category"
-                @click="() => deleteCategory(checkedCategories)"
-              >
-                <i class="icon-trash" />
-              </button>
-              <button
-                class="btn btn-warning btn-xs"
-                :disabled="checkedCategories.length < 2"
-                data-toggle="modal"
-                data-target="#mergeConfirmation"
-              >
-                <i
-                  class="icon-merge-rows"
-                  data-toggle="tooltip"
-                  title="Merge selected categories"
-                />
-              </button>
-            </div>
-          </div>
-          <button
-            class="btn btn-info btn-block"
-            :disabled="!currentCategoryFormValid"
-            @click="() => addCategory()"
-          >
-            <i class="icon-plus" /> Add Category
-          </button>
-        </div>
-      </div>
-      <div
-        v-if="!currentCategoryFormValid || !currentLabelsValid"
-        class="h-error-messages"
-      >
-        <p class="form-validation-error">
-          Please fix all errors to continue
-        </p>
-        <ul v-if="currentFormErrors.length > 0 || currentLabelingErrors.length > 0">
-          <li
-            v-for="error in currentFormErrors"
-            :key="error"
-            class="form-validation-error"
-          >
-            {{ error }}
-          </li>
-          <li
-            v-for="error in currentLabelingErrors"
-            :key="error"
-            class="form-validation-error"
-          >
-            {{ error }}
-          </li>
-        </ul>
-      </div>
-      <hr v-if="showLabelingContainer">
-      <button
-        v-if="showLabelingContainer"
-        class="btn btn-primary btn-block"
-        :disabled="!currentCategoryFormValid || !currentLabelsValid"
-        @click="beginTraining"
-      >
-        <i class="icon-star" /> Begin training
-      </button>
-    </div>
-    <active-learning-merge-confirmation
-      id="mergeConfirmation"
-      :callback="mergeCategory"
-      :category-name="currentCategoryLabel"
-      :fill-color="currentCategoryFillColor"
-      :selected-labels="selectedLabels"
+    <!-- Labeling Dialog -->
+    <active-learning-labeling
+      :image-names-by-id="imageNamesById"
+      :labeled-superpixel-counts="labeledSuperpixelCounts"
+      :current-form-errors="currentFormErrors"
+      :category-index="categoryIndex"
+      :pixelmap-rendered="pixelmapRendered"
+      :available-images="availableImages"
+      @combine-categories="combineCategoriesHandler"
+      @update="synchronizeCategories"
     />
+    <!-- Slide Image -->
     <div
       ref="map"
       class="h-setup-categories-map"
     />
+    <!-- Opacity Slider -->
     <annotation-opacity-control
       :style="{'width': '350px', 'right': '20px'}"
-      :fill-color="currentCategoryFillColor"
+      :category-index="categoryIndex"
       :overlay-layers="[overlayLayer]"
       :disabled="!overlayLayer"
     />
+    <!-- Information Panel -->
     <div :class="{'h-setup-categories-information': true, 'h-collapsed': !showInfoContainer}">
       <mouse-and-keyboard-controls
         v-if="showInfoContainer"
@@ -847,22 +478,6 @@ export default Vue.extend({
 </template>
 
 <style scoped>
-.h-labeling-container {
-    z-index: 1000;
-    position: absolute;
-    top: 5px;
-    left: 5px;
-    width: 400px;
-    display: flex;
-    flex-direction: column;
-    background-color: #fff;
-    border-radius: 5px;
-    box-shadow: 5px 5px 5px rgba(0, 0, 0, 0.5);
-    padding: 5px;
-    height: 90vh;
-    justify-content: space-between;
-}
-
 .h-collapsed {
     max-width: fit-content;
     height: auto;
@@ -873,33 +488,6 @@ export default Vue.extend({
     background-color: transparent;
     width: fit-content;
     height: fit-content;
-}
-
-.h-container-title {
-    display: flex;
-    background-color: #f6f6f6;
-    border-radius: 5px;
-    align-items: center;
-    min-height: 40px;
-    padding-right: 5px;
-}
-
-h4 {
-    margin: 10px auto;
-}
-
-.h-al-setup-categories {
-    border-radius: 5px;
-    height: 100%;
-}
-
-.h-al-image-select {
-    width: 100%;
-    padding: 5px 10px;
-}
-
-.h-al-image-select-new {
-    box-shadow: 0px 0px 5px 1px rgba(0, 127, 0, 0.5);
 }
 
 .h-setup-categories-information {
@@ -919,87 +507,4 @@ h4 {
     min-height: 100vh;
     border: 1px solid #f0f0f0;
 }
-
-.h-form-controls {
-    max-width: 550px;
-    display: flex;
-    flex-direction: column;
-}
-
-.h-category-form {
-   display: flex;
-   flex-direction: column;
-}
-
-td, th {
-    padding: 0px 2px;
-    text-align: center;
-}
-
-th:nth-child(2), td:nth-child(2) {
-    text-align: left;
-}
-
-tr:nth-child(even) {
-    background-color: #f0f0f0;
-}
-
-tr .editing-icons {
-    opacity: 0;
-}
-
-tr:hover .editing-icons {
-    opacity: 1;
-}
-
-.table {
-    margin-bottom: 5px;
-}
-
-.table-labels {
-    display: flex;
-    justify-content: space-between;
-    min-width: 100%;
-}
-
-.h-selected-row {
-    font-weight: bold;
-}
-
-.h-error-messages {
-    padding-top: 25px;
-    height: 100%;
-}
-
-.form-validation-error {
-    color: red;
-}
-
-.h-table-button {
-    padding: 0px;
-    background-color: transparent;
-    color: #888;
-}
-
-.condensed-color-picker >>> .input-group-addon {
-    padding: 0px;
-    border: none;
-    background-color: transparent;
-}
-
-.condensed-color-picker >>> .form-control {
-    display: none;
-}
-
-.category-input {
-    height: 20px;
-    padding: 0px 5px;
-}
-
-.h-remove-categories {
-    display: flex;
-    justify-content: flex-end;
-    margin-bottom: 5px;
-}
-
 </style>
